feat: store project in IndexedDB local storage for persistence
 - New IDBStoredProject - stores user files/code in IndexedDB
 - Load IDBStoredProject on start
 - Proper new project initialization
 - Abstract TreeView to support showing both IDBStoredProject and EMFS
 - Add project buttons on top bar

Details:
User's files are now stored primarily in an IndexedDB database,
instantiated for their project. Files are mirrored from a virtual
filesystem created in the IndexedDB(see IDBStoredProject) to
the virtual filesystem in the Emscripten SplashKit Module.
This mirroring processing is displayed in the FileView -
 - White = only in IDB
 - Semi Transparent White = only in EMFS
 - Green = mirrored
This allows transient files to be created while the program is running,
without having to save/store them in the IDB (or when exported as a zip).
They will simply be cleared on reload.