feat: vastly improved code execution: async + scope handling + UI changes
This commit makes two major changes, enabled by modifying the user's
input source code.
 - The input code is 'asyncified' so that it can be ran asynchronously.
   This lets the user create infinite while loops and other useful
   constructs without freezing the page.
 - The global variables (and functions/classes) declared in the input code
   are detected and manually lifted to directly set 'window'.
   This makes variable scoping behave more like expected in standard
   Javascript, and also allows us to run the code in "Use Strict" mode.
   We can also reset the global variables, to give the user a (mostly)
   clean slate, without destroying/recreating the entire
   execution environment.

These changes are heavily documented in
executionEnvironment_CodeProcessor.js

Because of this, there are also several UI changes - the user's code is
now in semi-generic 'blocks', and the actual running of the program
itself is in a seperate area. Programs can now be paused/resumed,
and blocks re-ran to hot-reload functions and variables. It is expected
that a program now have a 'main' function, that loops until program
termination, rather than the artifical Init/MainLoop seperation previously.